drivers: jesd204: move folder from iio
This change re-organizes the `drivers/iio/jesd204` into a `drivers/jesd204`
structure, with the intent of creating the JESD204 sub-system, and
integrating this code into it.
