+++ drivers/iio/Kconfig
@@ -81,7 +81,6 @@ source "drivers/iio/gyro/Kconfig"
 source "drivers/iio/health/Kconfig"
 source "drivers/iio/humidity/Kconfig"
 source "drivers/iio/imu/Kconfig"
-source "drivers/iio/jesd204/Kconfig"
 source "drivers/iio/light/Kconfig"
 source "drivers/iio/logic/Kconfig"
 source "drivers/iio/magnetometer/Kconfig"
