+++ drivers/iio/Makefile
@@ -27,7 +27,6 @@ obj-y += frequency/
 obj-y += health/
 obj-y += humidity/
 obj-y += imu/
-obj-y += jesd204/
 obj-y += light/
 obj-y += logic/
 obj-y += magnetometer/
