--- drivers/iio/jesd204/Makefile
@@ -1,6 +1,0 @@
-obj-$(CONFIG_AXI_ADXCVR) += axi_adxcvr.o axi_adxcvr_eyescan.o
-obj-$(CONFIG_AXI_JESD204B) += axi_jesd204b_v51.o axi_jesd204b_gt.o
-obj-$(CONFIG_AXI_JESD204_RX) += axi_jesd204_rx.o
-obj-$(CONFIG_AXI_JESD204_TX) += axi_jesd204_tx.o
-obj-$(CONFIG_XILINX_TRANSCEIVER) += xilinx_transceiver.o
-obj-$(CONFIG_ALTERA_ARRIA10_JESD204_PHY) += altera_adxcvr.o
