+++ drivers/jesd204/Makefile
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: GPL-2.0
+#
+# Makefile for the JESD204 core.
+#
+
+obj-$(CONFIG_AXI_JESD204B) += axi_jesd204b_v51.o axi_jesd204b_gt.o
+
+obj-y += linklayer/
+obj-y += transceiver/
