+++ drivers/jesd204/linklayer/Makefile
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: GPL-2.0
+#
+# Makefile for JESD204 Link Layer Peripheral drivers
+#
+
+# When adding new entries keep the list in alphabetical order
+obj-$(CONFIG_AXI_JESD204_RX) += axi_jesd204_rx.o
+obj-$(CONFIG_AXI_JESD204_TX) += axi_jesd204_tx.o
