+++ drivers/jesd204/transceiver/Kconfig
@@ -1,0 +1,19 @@
+#
+# JESD204 Transceiver drivers
+#
+
+menu "JESD204 Transceiver drivers"
+
+config ALTERA_ARRIA10_JESD204_PHY
+	tristate "Altera Arria10 JESD204 PHY Support"
+	depends on COMMON_CLK
+
+config AXI_ADXCVR
+	tristate "Analog Devices AXI ADXCVR PHY Support"
+	depends on COMMON_CLK
+	select XILINX_TRANSCEIVER
+
+config XILINX_TRANSCEIVER
+	tristate
+
+endmenu
