+++ drivers/jesd204/transceiver/Makefile
@@ -1,0 +1,9 @@
+# SPDX-License-Identifier: GPL-2.0
+#
+# Makefile for JESD204 Transceiver drivers
+#
+
+# When adding new entries keep the list in alphabetical order
+obj-$(CONFIG_ALTERA_ARRIA10_JESD204_PHY) += altera_adxcvr.o
+obj-$(CONFIG_AXI_ADXCVR) += axi_adxcvr.o axi_adxcvr_eyescan.o
+obj-$(CONFIG_XILINX_TRANSCEIVER) += xilinx_transceiver.o
